upload multi images

+++ src/main/java/com/benkitoucoders/ecommerce/services/ImageService.java
@@ -3,13 +3,21 @@ package com.benkitoucoders.ecommerce.services;
 import com.benkitoucoders.ecommerce.daos.ImageDao;
 import com.benkitoucoders.ecommerce.dtos.ImageDto;
 import com.benkitoucoders.ecommerce.entities.Image;
+import com.benkitoucoders.ecommerce.exceptions.EntityAlreadyExistsException;
 import com.benkitoucoders.ecommerce.exceptions.EntityNotFoundException;
 import com.benkitoucoders.ecommerce.mappers.ImageMapper;
 import com.benkitoucoders.ecommerce.services.inter.ImageServiceInter;
+import com.benkitoucoders.ecommerce.utils.Constants;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -22,8 +30,8 @@ public class ImageService implements ImageServiceInter {
     private final ImageMapper imageMapper;
 
     @Override
-    public List<ImageDto> getImagesByQuery(Long imageId, String imageUrl, Long productId) {
-        List<ImageDto> imageDtoList = imageDao.findImagesByQuery(imageId, imageUrl, productId);
+    public List<ImageDto> getImagesByQuery(Long imageId, String imageName, String imageType, String imageFilePath, Long productId) {
+        List<ImageDto> imageDtoList = imageDao.findImagesByQuery(imageId, imageName, imageType, imageFilePath, productId);
         return Optional.ofNullable(imageDtoList)
                 .filter(list -> !list.isEmpty())
                 .orElse(Collections.emptyList());
@@ -31,7 +39,7 @@ public class ImageService implements ImageServiceInter {
 
     @Override
     public ImageDto getImageById(Long id) {
-        List<ImageDto> imageDtoList = imageDao.findImagesByQuery(id, null, null);
+        List<ImageDto> imageDtoList = imageDao.findImagesByQuery(id, null, null, null, null);
         return imageDtoList.stream()
                 .findFirst()
                 .orElseThrow(() -> new EntityNotFoundException(String.format("There is no image with the ID %d.", id)));
@@ -54,7 +62,67 @@ public class ImageService implements ImageServiceInter {
 
     @Override
     public void deleteImageById(Long id) {
-        getImageById(id);
-        imageDao.deleteById(id);
+        ImageDto image = getImageById(id); // Make sure this method retrieves the image entity
+
+        System.out.println(image);
+
+        // Delete the image file from the file system
+        String filePath = image.getFilePath();
+
+        System.out.println(filePath);
+
+        File imageFile = new File(filePath);
+
+        if (imageFile.exists()) {
+            if (imageFile.delete()) {
+                imageDao.deleteById(id);
+            } else {
+                throw new RuntimeException("Failed to delete image file");
+            }
+        } else {
+            // If the file doesn't exist, only delete the entity from the database
+            imageDao.deleteById(id);
+        }
     }
+
+
+
+    @Override
+    public String uploadImageToFileSystem(MultipartFile file, Long productId) throws IOException {
+        String folderPath = Constants.FOLDER_PATH + "product_" + productId + "/";
+        String filePath = folderPath + file.getOriginalFilename();
+
+        Optional<Image> existingImage = imageDao.findByFilePath(filePath);
+        existingImage.ifPresent(image -> {
+            throw new EntityAlreadyExistsException(String.format("This image with name %s already exists", file.getOriginalFilename()));
+        });
+
+        Image image = imageDao.save(
+                Image.builder()
+                        .name(file.getOriginalFilename())
+                        .type(file.getContentType())
+                        .filePath(filePath).build());
+
+        // Create the product's folder if it doesn't exist
+        Path path = Paths.get(folderPath);
+        if (!Files.exists(path)) {
+            Files.createDirectories(path);
+        }
+
+        if (image != null) {
+            file.transferTo(new File(filePath));
+            return "file uploaded successfully: " + filePath;
+        } else {
+            return "Failed to save image entity.";
+        }
+    }
+
+
+    @Override
+    public void uploadImagesToFileSystem(List<MultipartFile> images, Long productId) throws IOException {
+        for (MultipartFile image : images) {
+            uploadImageToFileSystem(image, productId);
+        }
+    }
+
 }
